mail: Create the thread index as soon as possible [xopgi].
This has to be done cause the `create` method of threads may entail the
sending of notifications.  Such notifications won't have the index built
at the time because the index is built too late for them.

Since the index needs to know the res_id, we have to patch the Odoo's
method.

Watch for conflicts.