[bus] Reduce polling when the page is not visible.
Whenever a polling request ended without error it immediately started a
new request.  If the user was not really paying attention to the page
this wasted resources.

This patch changes this behavior.  If the page is visible at the time
the polling request ends, it will behave roughly as before.  If the page
is hidden, then it will wait for the page to become visible for a
limited amount of time before polling again.  If the browser doesn't
support the visibility API it will behave as if the page is visible at
all times.