[http] Ignore etag weakness for static assets.
Some (nginx) proxies will weaken the etags if they perform a transformation like
gzipping.  But then browser will send a weak etag that the proxy will pass as
given (cause is an end-to-end header) and the Odoo will not match it.

Since Odoo always produces strong tags, this patch wraps the WSGI application in
a middleware that simply remove the "W/" for etags.