[celeryapp] Put static arguments first.
Static arguments are those not related to the environment but to the
code.  These are the model and method the deferred is bound to.  These
don't change dynamically when the program is running.  By putting them
first, we ease the construction of partials over the task and pave the
way for future work that involve chaining of tasks.