[celeryapp] Fix event-handling for background jobs
The bus will now spawn a thread (greenlet) for watching events from
Celery Workers, specifically `task-succeeded` and `task-failed` events.

This allows to send the status success/failure notification outside the
task itself. DB errors (closed cursors, etc) in the task won't affect
the notification of its status.

We could even improve on the bus by avoiding the DB altogether.  But a
balance must be achieved.

We have also introduced a way to notify about the user when he or she is
waiting for the job to complete.