jobs: Revert reports of success back to 'normal'.
Events monitoring seems to be designed to monitor all jobs from all queues,
so they get more job than needed.  Furthermore the action of notification
about a task completion (although fallible) is also an application-level
action.

This commit reinstates reporting of success/failure at the application
level and not via event-monitoring.  Anyways reports are also executed as a
celery task so that failure of reporting does rollback any change actually
done by the task.

Another benefit is that we don't need to spawn the gevent process to get
report notifications, and so this is more robust to changes.