[FIX] im_chat: Don't change controller arguments - can lead to failures.
This solves an edge case when other (custom) addons inherit from the
bus.Controller and override the _poll() method.

If any of those addons may fail in a way the Odoo server retries the
request (e.g. transactions in the DB), then `channels` gets reused in
the `poll` controller and the assertion of channels being a string
fails: "bus.Bus only string channels are allowed."

Solution proposed: don't modify the channels list, but create a new one.