http: Allow several domains for the session cookie.
When serving the same Odoo instance over several domains we must allow
the session cookie to be secured for that domain.

In this case we select the first match from the request's host and the
allowed domains.  This means the sessions are not cross-domains.

If any of the allowed domains starts with a dot '.' the sub-domains are
allowed.