jobs: Avoid generating too many queues and the backend.
Queues were piling up.

We don't a have a way to purge old tasks' results from the backend.
Furthermore, this backend was only needed to check the status of job in
web_celery.  Now there's a small chance that by the time the client
receives the WAIT_FOR_TASK action, the task be already done.

TODO: Review design.