jobs: Leave a single queue.
Only time will tell which queues we'll need.  So why bother having so many of
them.

Also we can generalize the workers needed in the 'server'.  So cut them as
well.

This work unveiled an error, the beat worker was not being killed.  Since
we're not actively using the beat worker this was not noticed.  Fixed.

For the time being, we're leaving the `HighPriorityDeferredType`,
`LowPriorityDeferredType`, `LowPriorityDeferred`, and `HighPriorityDeferred`
names as aliases to the Deferred.