workflow: Improve integration with celery jobs.
If an error (like an AssertionError or ValidationError) should happen
while evaluating a workflow step, Celery tries to reconstruct the
whole traceback, but copies it so it can be pass to pickle (see
billiard/einfo.py).

While copying the traceback, it will reach the safe_eval that has the
workflow's Environment as the globals.  The __getitem__ will try to
reach the Odoo's api Environment within the thread's locals data and
may raise an AttributeError.

So if the workflow's Enviroment is to behave like a normal dict, the
AttributeError must be converted to a KeyError.