jobs, bus: Use a different cursor for sending bus messages.
This was discovered by analyzing why celery jobs would not rollback
the entire transaction after an error.

Since celery jobs may use the bus to send messages to clients watching
the progress of the job, the cursor is actually being committed and
all the changes before the error (actually before the last bus
message) are kept.

In normal HTTP request this does not happen (yet) because no single
use case but the chat is sending a bus message.

The true solution is, thus, to make the bus use another cursor to
clearly separate messages from normal data.

Notice the explicit commit is no longer necessary: we create all
messages, garbage-collect old ones, commit (by exiting the with block)
and then notify clients.