jobs: Allow to create more queues.
We have been witnessed bombs of long-lasting jobs that created a head-of-line
blocking effect in the default queue.  For instance, the periodic CDR cycle
has been put in the backlog behind a few jobs that took the whole soft-time
limit.

The idea is to have a few queues and be able to run an equal (or greater)
number of dedicated workers.

For the time being, we'll keep forking the worker for the default queue.  But
soon enough we won't fork the worker by default; a new command `openerp
worker` will run the worker.

This second decision has another rationale: the server was getting much more
complex, we're not entirely sure if a bug about the WorkerController not
reaping the PoolWorker processes is due to the way the Odoo's Workers handle
the SIGCHLD signal.  But we should not change that without fully understanding
the consequences.  So a better approach is to have the workers managed by
supervisor or other daemon.