jobs: New API for Deferred.
We need the Deferred to support calling without cr, uid and the like to be
able to migrate to Odoo 10.

The new API for Deferred has been totally simplified.  The old API is kept for
backward compatibility.

You are urged to migrate to the new API as soon as possible.