celery: Remove sentry's handlers.
Registering those signals creates a lot events in the Sentry which are
ultimately logged elsewhere.  Specially because we're funneling all jobs into
tasks and the tasks do the background jobs.

Probably this registry stuff is appropriate and we must review it.  But, for
the time being, let's keep our sentry saner.