jobs: Fix retries not being properly tracked.
Retries have a different job UUID and, thus, if you're subscribed to the
channel of the *original* job ID you won't get any notifications from the
retried job.  So we keep the original UUID so that we can send notifications
and reports properly.

We also noticed that if the default queue is clogged with many tasks finished
or failed tasks won't receive the report until the the report-task gets a slot
in a worker.  The quickest fix for this is to create a separate queue for the
notifications.   This way you may create a separate worker for it (they don't
do much).

On the same account, maximum retries were not being properly handle in any of
the tasks, and the default retry delay was too high (3 minutes).