merchise mail: Recover all the data from a message.
In several of our addons we modify the 'message_format' and/or 'read' of the
mail.message to include more values to the message.   Odoo, ignores and
removes anything that is not the fields before the lines we touch in this
commit.

We recover all that data so that addons can use it.