jobs: Improve detection of hard timeouts.
If any task ignores the SoftTimeErrorLimit, the job still may get
forcibly killed.  Even though that keeps our server happy, it hurts
our ability to inform the inpatient user about the failure of the job.

This hack allows us to detect the hard timeout and report it back to
the user.