jobs: Improve error handling.
The `MaybeRecords` (a name which I know is awfully wrong) context manager
needs to load the DB environment and they may raise DB OperationalError.
Specially, under high load this means that some tasks may fail even without
getting a chance to start.

We up-lift the ``try``..``except`` so that it wraps the entire task.

Also we not that `_report_failure_for_request` was catching all types of
errors.  Only user-level errors (Exception and below) should be ignored in
this case.