jobs: Add support to disambiguate `until_timeout`.
When `until_timeout` terminated we could not known if it was because the
iterator was fully consumed or because of a timeout.

We add a `on_timeout` callback so that we can be flagged about termination
because a timeout.  In addition we provide an EventCounter that can be used to
produce instance for `on_timeout`.

The canonical example would be:

   timed_out = EventCounter()
   for _ in until_timeout(iterator, on_timeout=timed_out):
      pass

   if timed_out:
      pass