jobs: Add some support for nested calls to `until_timeout`.
We had to remove the entire bounds-based implementation.  `until_errors` does
not mix well with other boundary conditions.  This have to be rethought in
xoutil.

Nevertheless, the implementation is now much more simple to understand.  Which
is another argument to the need of a review of `xoutil.bounds`.  It seems that
we need something better (monads or arrows?) in order for `xoutil.bounds` to
be general enough.

The EventCounter had to be extended to support chaining of counters.

We included tests in the 'merchise-tests/' directory.  The name of the
directory was chosen to avoid possible future clashes if Odoo ever includes a
top-level 'tests/' directory.  `tox` will default to find tests in such
directory.

This change removes the need of the comment introduced in ec72247ecdc.

However, as demonstrated in the tests, context and iterators does not mix
well.  The stacking of a context follows a path within the tree and does not
preserve the tree itself so we can't make any warrants about the `on_timeout`
in such cases.