account: Remove dead code.
Notice that the only effect this code has is to produce a ``.write()``.  That
write only occurs only if any of the variables ``code`` or ``amount`` were
updated to something different than 0.  Those variables are only touch there's
a matching key inside ``account2`` (which is always the empty dict) or if
there's a matching account id inside ``account`` (which is always empty as
well).  So, the code does nothing.