models: Send the stack when 'ensure_one' fails.
Our integration with Sentry avoid logging `except_orm` errors since they are
used for both user errors and system errors.

This commit does not attempt to fix the real issue of providing a single
purpose to `except_orm`, since then we'd have to review the whole code base.
Simple we force the log along with the stack.