celery: Remove invalid assumption.
The only mention of 'multi_process' inside addons and openerp/addons is in the
'bus' and during initialization.  The issues with report and other addons
seems to be unrelated to this flag.

Besides the ``RegistryManager.check_registry_signaling()`` method bails if
multi_process is False: there's no point to do that in that case.