product: Split `_compute_price_rule` method so that we hack its internals.
Originally, the method `_compute_price_rule()` must select a single rule per
product and compute the price, it does so by traversing the pricelist-rules
tree and finding the 'appropriate' rule.

We split the entire process in:

- Find the rule that matches the date, product and quantity.  Method
  `_get_rule`.

- Compute the base price for the rule found (this may entail another
  pricelist).  Method `get_base_price`.

- Compute the final price by passing the base price to the formula 'base * X +
  Z' or the python code.  `get_final_price`.

This way we can change how the process is done for complex price computations.
e.g: We need to compute the price of a single sale order line for several
days (hotel rooms).

At the same time, we optimize the SQL by selecting rules that match the
minimum required quantity requirement (this was previously done in Python).