server: Try to kill the worker from within.
After the real-time limit is hit by a worker, we first signal the worker with
SIGXCPU (?) and after a tiny amount of time we actually kill the worker if it
has not finished.

Signalling the worker allows to raise an exception within it and inspect the
slow request in the Sentry.