edi: Avoid test to fail because of mail thread indexes.
Our xopgi additions in `addon/mail` introduce a thread index in ir.model.data
for all thread records.  So, when importing a new record from EDI and
generating it's EDI information again it may end with another __id.

Is this an issue for EDI?  I don't know, but since EDI is removed in Odoo 9+,
it doesn't really matter.

Fixes mercurio/xhg#11.