signals: Remove signals module.
Keeping the `xoeuf.signals`:mod: module in sync with this is time consuming.
The only reason for having the signals here is we use them in this repo (Odoo
should not depend on xoeuf).  But that's not the case anymore.

Please forgive us and use `xoeuf.signals`:mod: which is cool and improving.