web_celery: Adapt to Odoo 12.
The web client was heavily changed in Odoo 12.

Fortunately, the new bus service can be used as a singleton because
addChannel/deleteChannel do restart the longpolling connection; so we can use
the bus_service directly.

The action no longer requires us to insert in the DOM.  The AbstractAction
willStart method inserts the action's controller in the DOM.  So we only need
to hook into attach/detach callbacks.

There's still some issues.  When there's an error, the previous screen is not
restored properly.