[FIX] fetchmail: Ensure mails are processed entirely or not.
Some malformed emails may contain NUL characters in their bodies.  If the
message contains attachments Odoo creates the attachments (and possibly a new
object), but fails to create mail.message.

Furthermore the mail is left in the INBOX which results in attachments piling
up until we manually delete the bad email from the INBOX.

This commits avoids creating the objects in the DB unless everything goes OK.
It doesn't help you from having to manually delete bad emails in the INBOX.