Store completely_defined value.
== Rationale ==

Listing the pricing steps in production, for moderately complex pricing
programs, is taking too long.

The issue is that the field 'completely_defined' is being shown in the listing
and needs to be computed for all the steps in the list.

== The Implementation Explanation ==

While doing this I found another issue with the fields.Enumeration which is
not easily overcome.

Odoo launches the recomputation of 'completely_defined' at the time when the
addon 'xhg_autrement_prices' is being loaded into the registry.  But the
enumerations we use here are built dynamically, and some of their values DON'T
exist yet when this recomputation happen and it fails with a ValueError
'Invalid member %s of enumeration'.

Odoo's fields.Selection doesn't suffer from this issue because the wrong value
in the cache is a simple string which they can put and validate later.

But we cannot prime the cache with a False value because then once the
recomputation is done it won't be triggered again.

The solution is to do the recomputation ourselves after all addons have been
loaded (see the hacks.py module).