http: Reduce the maximum amount of entries in the metrics.
The idea is to reduce the likelihood of have too-old entries when Prometheus
asks for the logs.

Since each WorkerHTTP has it's own deque, each time Prometheus fetches the
entries it actually receives only a partial view of the system.  If we keep
too many entries, Prometheus could receive data that is too old (from a worker
that it didn't reach for a long time but was keeping entries).

The idea is that entries should be fresh, at the expense of loosing data
points.

Or we must keep the entries in a central storage.

Our current demand is less than 20 reqps, divided by 8 workers (2.5 reqps per
worker); so 256 entries is about 1 min of data.