Invalidate Odoo cache while ensure mail threads indexes.
This issue was brought to my attention while investigating the performance
issue in steps.

In the presence of computed fields with inverse the method `create()` would
invoke the inverse, which most likely writes to the DB calling `_ensure_index`
before it's has been called by mail.threads' method `create`.

At this moment the thread_index in the cache is False and so a new index is
created but the cache is not invalidated.  This happens several times (one per
inverse call) and we may create several references to the same object.

So we need to invalidate the cache after creating the indexes.

Before this commit, creating a new DB and installing xhg_autrement_prices
results in the creation of up to 4 indexes per step definition:

    prices=# WITH tmp AS (SELECT count(*) as times, res_id, model
	                      FROM ir_model_data WHERE module='__mail_threads__'
						  GROUP BY res_id, model ORDER BY times, model)
             SELECT * FROM tmp WHERE times > 2;

	 times | res_id |                 model
	-------+--------+---------------------------------------
		 3 |      2 | xhg.autrement.pricing.step.definition
		 3 |      3 | xhg.autrement.pricing.step.definition
		 3 |      1 | xhg.autrement.pricing.step.definition
		 3 |      5 | xhg.autrement.pricing.step.definition
		 4 |      4 | xhg.autrement.pricing.step.definition
		 4 |      6 | xhg.autrement.pricing.step.definition

With this commit:

    prices=# WITH tmp AS (SELECT count(*) as times, res_id, model
	                      FROM ir_model_data WHERE module='__mail_threads__'
						  GROUP BY res_id, model ORDER BY times, model)
             SELECT * FROM tmp WHERE times > 0;

	 times | res_id |                 model
	-------+--------+---------------------------------------
		 1 |      1 | mail.channel
		 1 |      2 | mail.channel
		 1 |      3 | mail.channel
		 1 |      4 | mail.channel
		 1 |      2 | res.partner
		 1 |      5 | xhg.autrement.pricing.step.definition
		 1 |      3 | xhg.autrement.pricing.step.definition
		 1 |      2 | xhg.autrement.pricing.step.definition
		 1 |      1 | xhg.autrement.pricing.step.definition
		 1 |      4 | xhg.autrement.pricing.step.definition
		 1 |      6 | xhg.autrement.pricing.step.definition