Start release 191.
The problem: Putting our release in the addon's versions creates the problem
with our rolling-releases schema that we can't bump the version till late
in the development cycle.

But then migrations that reach master are not run in the 'development' and
'staging' branches, which causes some errors in those environments.

The solution:

Use a single-number for our releases and put it as the second number of the
Odoo version.  Our next release was named 1.9.1, so we mark the next release
as 12.191.  Our numbers will increased by one: 192, 193, ...

Addons can have a numbering of their own and bump the version whenever they
want.