job: Run Deferred jobs in-process when running tests.
Testing code that's supposed to run a celery job is hard (and not done in our
CI pipeline) because it requires several integrated pieces of software (the
broker, the worker, polling for results, etc) and we don't have the resources
to deploy such a beast.

There are two strategies:

- Making our Celery jobs just a very thin layer on top of regular methods and
  call those methods in our tests.

- Allow calls to Deferred while running tests in a way that those calls are
  run in-process and not actually transferred to a Celery worker.

IMPORTANT: If you can, DO it the first way.  Methods calling Deferred methods
should usually return a Web Client action, and therefore they should be
separated anyways from the code of the actual job.