web_celery: Revamp the celery jobs JS notification system.
Now we provide a lower-level widget which is not triggered by a Client
Action.  Using a client action has the undesired side-effect of dropping the
current controller and the work-area is cleared.

By using a lower-level mechanism (hooking the AJAX calls) we can display our
notification widgets without loosing the controller.

We introduce a Web Client Service to coordinate the whole thing.