job: Actually signal the registry change after a task.
Since commit merchise/odoo@b59318ec12a1ce4e59de6d7bcde7146545b6c044 the
registry can be replaced within the boundaries of a task, so signalling must
be done in the outer layers of the framework (ir_cron, http, and now jobs).

I tried first doing just:

     with registry.manage_changes():
	   ...

But that didn't work properly (the web-server process was not notified of the
registry being changed).  Installing a new module flags *another* registry and
not the same we started with.