fetchmail: Ensure mails are processed entirely or not
While fetching emails Odoo creates the attachments (ir.attachment) and
possibly the target object (eg. crm.lead) before creating the mail.message. If
the mail.message fails to be created the attachments are left in the DB.

Furthermore, the mail is left in the INBOX which (in the case a POP server)
results in attachments piling up until we manually delete the email from the
INBOX.

This commits avoids creating the objects in the DB unless everything goes
OK. It doesn't help you from having to manually delete bad emails in the
INBOX.

The issue has been witnessed in DBs with 3rd party addons. But Odoo should be
robust in those cases. We include that simulate a message insertion failure.

This was sent to Odoo in PR https://github.com/odoo/odoo/pull/31157 but it
hasn't been merged.