jobs: Don't send the valuemax at the end of iter_and_report.
Otherwise the two-stages implementation will fail.  It's best to let either
client of `iter_and_report` or the web client to *complete* that last chunk.

This was here because if the count of items is not a multiple of rate, then
the last chunk was not sent reported to the user.  Now simply send out-of-rate
the progress report.