Improve logging of QWeb related exceptions.
This will allow Sentry to receive more meaningful data.  See issue #852.

The test suite makes some HTTP request which are then logged to the stderr,
and we use stderr inspectption to know if some thing failed.  So if we log the
exception where it's being caught the tests in test_website are reported
broken.  Currently the tests passed because the use a MockRequest with a
customized _handled_exception.

This partially reverts commit 54977008194da59bc01df7495ff6a331cc989200.