http: Reduce the cache time and remove etag middleware.
The STATIC_CACHE was set when we used our modifications of assets via HTTP/2
which ensured the URL changed when the file (CSS/JS) did.

However, we removed that thing long time ago and forgot about this cache.
Furthermore this cache also affects XML files which are lazily loaded when
widgets declare xmlDependencies.

We've been witnessing some issues with caches not being invalidated, and
such.  So let's get a little bit closer to Odoo's standard behavior.