web_celery: Allow to cancel background jobs.
Cancelling background could be necessary in three broad scenarios:

- The user explicitly cancels the background job

- The user abandons the page for which the background job is relevant

- The user performs an action that implies the cancellation of the job.

We added a couple of functions:

- `odoo.jobs.terminate_task_with_env`
- `odoo.jobs.terminate_task`

Both try to provide a different level of security to avoid malicious
cancellation of background jobs from unauthorized parties.

This commit also starts a refactor of 'odoo.jobs' to create small classes to
deal with the signature, task *request* (we call it record to avoid confusion
with the celery Request), etc.

Many of the 'web_celery' JS was documented or its documentation improved.