Report progress of price table generation more often.
A fix rate of times doesn't really work when there are thousands of prices to
be computed.  We would pace at rate of hundreds of prices and that would
appear to the user as a stalled process.

For instance, a table computing about 30000 prices (we have such cases) would
issue the first report at the 1500th item.  With the current algorithm it
takes to long and the user sees no progress.

The new odoo.jobs.ReportRate object allows a finer control about the rate at
which report are acceptable: you may control for not-too-fast, not-too-slow
, not-too-few and not-too-many.

For the price tables we have settled to a minrate of 1 item (at least one
price has to be computed since the last report) but then we also limit the
rate 1 report per 5 seconds, but allow it to go up to 8 seconds before
reporting.  This still is perceived as slow (and slow it is), but it won't
leave the user thinking there's no progress at all.