Propose an API for the staged reporting of background jobs.
There are background jobs which are performed in stages.  Each stage may take
a variable amount of time and their overall progress values cannot be
estimated up-front.  For instance, computing the estimate of the main job
could be regarded itself as stage.

Possible application: creating the price tables.

We've include those changes in the commit as an example, even though the UI
does nothing with them.