Reduce the likelihood of failing while reporting a job failure.
When the arguments to the exceptions are not JSON serializable kombu can't
pass the message and fails.

This commits handles two possible cases:

- Retry reporting failures when the error itself can't be serialized.   This
  unlikely, but it's safe to do.

- Reduce the likelihood of failures by stripping the arguments from the
  report.

Reporting the failure is vital for the UI doesn't get staled whilst the
background job isn't anymore.

This commit takes care of the *reporting* issue described in #1192.  But we
still need to avoid the underling KeyError.