Ensure that every past stages reach the 100% of progress.
In some cases the server doesn't send the 100% of an stage we the UI leaves a
gap between stages which is not desirable.  In other cases we receive the
messages out of order and we may advance a later stage before advancing an
earlier one.

The treatment of staged was split into mixins StagedProgressBarMixin and
CeleryProgressBarMixin.  We remove the former to keep all this in a single
unit of code.