Don't track completely changes during Registry initialization.
We don't care about those changes usually, because the only steps in data are
two UndefinedProcedure which are never partially defined.

But our steps in demo data which are used in tests are indeed partially and
completely defined.  If we allow the Deferred to try to issue the job, the
tests would hang waiting for a connection to the redis server.