merchise: Remove mail's xopgi extensions.
These extensions were added in our fork long before Odoo 9 corrected the
`_inherit` algorithm which leaved some models without our extensions (in
`xopgi.mail`).

Furthermore, the index feature is only required when and only when
`xopgi_thread_address` is installed and emails are being sent/received from a
thread.  Now, we can probably implement the required indexing feature and not
in Odoo.

This commit fixes the failure:

    Traceback (most recent call last):
      File "/builds/merchise/odoo/odoo/tests/common.py", line 297, in assertQueryCount
        self.fail(msg % (login, count, expected, funcname, filename, linenum))
    AssertionError: Query count more than expected for user __system__: 2127 > 846 in test_leave_whole_company_07 at hr_holidays/tests/test_company_leave.py:320

Which reveals that indexing is a performance hog if done carelessly.