web: Make the JS tests 'notification_service_tests' pass.
The tests seem to be crafted in a way that depends too much on the ordering of
attributes in qweb.  I think this is failing for us because we use Python 3.8
and the QWeb implementation relies on implicit ordering-semantics that
changed.

The order of attributes is not relevant, though, so I'm just make them match.