Fix issue while routing emails forwarded from Gmail.
Gmail appends the header 'Delivered-To' when he receives the message and we do
the same when our server receives it.  This means that 'Delivered-To' can
appear multiple times in the message and we would get a space separated list
of addresses which 'tools.email_split' doesn't really understand.

This caused messages that were forwarded from gtcsrrf@gmail.com not be
properly.